pyproject.toml
[project]
name = "stock-analyzer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
	"pandas>=3.0.2",
	"python-dotenv>=1.2.2",
	"requests>=2.33.1",
	"yfinance>=1.3.0",
]

[tool.ruff]
# Target version for your Python environment
target-version = "py311"

[tool.ruff.format]
indent-style = "tab"
quote-style = "double"

[tool.ruff.lint]
# Enable Pyflakes (F), Error (E), and Implicit String Concatenation (ISC)
select = ["F", "E", "ISC", "I"]
# E501 is the "Line too long" rule. Adding it here kills the warning.
ignore = ["E501"]

[tool.ruff.lint.flake8-implicit-str-concat]
# Ensure it catches multi-line implicit concatenations
allow-multiline = false

[lint.isort]
combine-as-imports = true

[dependency-groups]
dev = [
    "coverage>=7.13.5",
    "openbb>=4.7.1",
    "openbb-yfinance>=1.6.2",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.11",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
